+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased Changes
+
+- Store attachments in `files/attachments/`
+  [#4](https://github.com/deltachat-bot/deltawoot/issues/4)
 
 ## 1.0.0